refactor(services/exporter): move function to service
