+++ crates/services/exporter/src/lib.rs
@@ -11,15 +11,15 @@ use database_models::{
     seen, user_to_entity, workout,
 };
 use database_utils::{
-    entity_in_collections, get_review_export_item, item_reviews, user_measurements_list,
-    workout_details,
+    entity_in_collections, item_reviews, user_measurements_list, workout_details,
 };
 use enums::EntityLot;
 use file_storage_service::FileStorageService;
 use fitness_models::UserMeasurementsListInput;
 use media_models::{
-    ImportOrExportExerciseItem, ImportOrExportMediaGroupItem, ImportOrExportMediaItem,
-    ImportOrExportMediaItemSeen, ImportOrExportPersonItem,
+    ImportOrExportExerciseItem, ImportOrExportItemRating, ImportOrExportItemReview,
+    ImportOrExportMediaGroupItem, ImportOrExportMediaItem, ImportOrExportMediaItemSeen,
+    ImportOrExportPersonItem, ReviewItem,
 };
 use nanoid::nanoid;
 use reqwest::{
@@ -228,7 +228,7 @@ impl ExporterService {
             let reviews = item_reviews(&self.db, user_id, &m.id, EntityLot::Metadata)
                 .await?
                 .into_iter()
-                .map(get_review_export_item)
+                .map(|r| self.get_review_export_item(r))
                 .collect();
             let collections = entity_in_collections(&self.db, user_id, &m.id, EntityLot::Metadata)
                 .await?
@@ -270,7 +270,7 @@ impl ExporterService {
             let reviews = item_reviews(&self.db, user_id, &m.id, EntityLot::MetadataGroup)
                 .await?
                 .into_iter()
-                .map(get_review_export_item)
+                .map(|r| self.get_review_export_item(r))
                 .collect();
             let collections =
                 entity_in_collections(&self.db, user_id, &m.id, EntityLot::MetadataGroup)
@@ -312,7 +312,7 @@ impl ExporterService {
             let reviews = item_reviews(&self.db, user_id, &p.id, EntityLot::Person)
                 .await?
                 .into_iter()
-                .map(get_review_export_item)
+                .map(|r| self.get_review_export_item(r))
                 .collect();
             let collections = entity_in_collections(&self.db, user_id, &p.id, EntityLot::Person)
                 .await?
@@ -394,7 +394,7 @@ impl ExporterService {
             let reviews = item_reviews(&self.db, user_id, &e.id, EntityLot::Exercise)
                 .await?
                 .into_iter()
-                .map(get_review_export_item)
+                .map(|r| self.get_review_export_item(r))
                 .collect();
             let collections = entity_in_collections(&self.db, user_id, &e.id, EntityLot::Exercise)
                 .await?
@@ -409,5 +409,33 @@ impl ExporterService {
             writer.serialize_value(&exp).unwrap();
         }
         Ok(())
+    }
+
+    fn get_review_export_item(&self, rev: ReviewItem) -> ImportOrExportItemRating {
+        let (show_season_number, show_episode_number) = match rev.show_extra_information {
+            Some(d) => (Some(d.season), Some(d.episode)),
+            None => (None, None),
+        };
+        let podcast_episode_number = rev.podcast_extra_information.map(|d| d.episode);
+        let anime_episode_number = rev.anime_extra_information.and_then(|d| d.episode);
+        let manga_chapter_number = rev.manga_extra_information.and_then(|d| d.chapter);
+        ImportOrExportItemRating {
+            review: Some(ImportOrExportItemReview {
+                visibility: Some(rev.visibility),
+                date: Some(rev.posted_on),
+                spoiler: Some(rev.is_spoiler),
+                text: rev.text_original,
+            }),
+            rating: rev.rating,
+            show_season_number,
+            show_episode_number,
+            podcast_episode_number,
+            anime_episode_number,
+            manga_chapter_number,
+            comments: match rev.comments.is_empty() {
+                true => None,
+                false => Some(rev.comments),
+            },
+        }
     }
 }
