+++ crates/utils/database/src/lib.rs
@@ -24,9 +24,7 @@ use file_storage_service::FileStorageService;
 use fitness_models::UserMeasurementsListInput;
 use itertools::Itertools;
 use markdown::to_html as markdown_to_html;
-use media_models::{
-    CreateOrUpdateCollectionInput, ImportOrExportItemRating, ImportOrExportItemReview, ReviewItem,
-};
+use media_models::{CreateOrUpdateCollectionInput, ReviewItem};
 use migrations::AliasedCollectionToEntity;
 use rust_decimal_macros::dec;
 use sea_orm::{
@@ -155,34 +153,6 @@ pub async fn entity_in_collections(
         entity_in_collections_with_collection_to_entity_ids(db, user_id, entity_id, entity_lot)
             .await?;
     Ok(eic.into_iter().map(|(c, _)| c).collect_vec())
-}
-
-pub fn get_review_export_item(rev: ReviewItem) -> ImportOrExportItemRating {
-    let (show_season_number, show_episode_number) = match rev.show_extra_information {
-        Some(d) => (Some(d.season), Some(d.episode)),
-        None => (None, None),
-    };
-    let podcast_episode_number = rev.podcast_extra_information.map(|d| d.episode);
-    let anime_episode_number = rev.anime_extra_information.and_then(|d| d.episode);
-    let manga_chapter_number = rev.manga_extra_information.and_then(|d| d.chapter);
-    ImportOrExportItemRating {
-        review: Some(ImportOrExportItemReview {
-            visibility: Some(rev.visibility),
-            date: Some(rev.posted_on),
-            spoiler: Some(rev.is_spoiler),
-            text: rev.text_original,
-        }),
-        rating: rev.rating,
-        show_season_number,
-        show_episode_number,
-        podcast_episode_number,
-        anime_episode_number,
-        manga_chapter_number,
-        comments: match rev.comments.is_empty() {
-            true => None,
-            false => Some(rev.comments),
-        },
-    }
 }
 
 pub async fn workout_details(
